Add a new project.
This is a stripped down version of the orchaed project which includes
I2C, SPI, flash and radio support, as well as the OLED display driver
and some of the ADC and GPIO modules. This project runs on the Freedom
KL019023 board and should serve as a fresh starting point.

If built correctly, the image should have 12688 bytes of RAM free. (Most
of the remaining RAM use seems to be from data structures that go in
the BSS.)

The flash code has been modified not to use the RAM relocation trick
in order to conserve the memory. We only need to relocate the flash
comamnd routine to RAM when trying to update a portion of the flash
which contains code, which is not anticipated at this time.