Reclaim a little more memory.
Avoid the use of rand() in the cmd-flash.c and radio.c code (the latter is
actually commented out right now). This reclaims an additional 8 bytes
of RAM. For some reason the rand() function in libc-nano uses malloc(),
which results in the use of a few additional bytes of BSS space.

Also, set the Makefile to use -std=c99 instead of -std=gnu99. We should
avoid GCC extensions. Ideally I would like to add -ansi -pedantic and
-Wconversion too, but that evokes a bunch of warnings from the graphics
library code, which I'm not sure I can fix because they're in a submodule.