Get CPU running at 48MHz, fix radioDump() command
The Freescale Freedom board uses FEI mode, which uses the KW01's internal
32KHz oscillator. At power up, the default MCG settings result in a 21MHz
CPU clock. We want it to be 48MHz, but for some reason ChibiOS doesn't
support this when FEI mode is selected. The hal_lld.c code has been modified
to set the MCG DRS and DMX32 bits as well as the OUTDIV1 and OUTDIV2 divisors
so that we now actually have a 48MHz CPU clock and 24MHz bus clock. Also,
the board.h file has been updated to indicate the new system clock value.

The radioDump() command didn't seem to be genrating so it's been corrected
(I think) to actually show all the internal radio registers.