Clean up new radio module
This commit adds documentation and copyright blocks to the radio module,
and cleans up the code a little.

The radio_select()/radio_unselect() functions have been removed since the
extra level of indirection really isn't needed. The functions now take
a RADIODriver handle in keeping with the convention of the previous code.
The driver now uses a single RADIODriver structure to manage all its state,
including a pointer to the SPI interface.