Enable support for LTO
Newer versions of GCC support link-time optimization. When enabled, it
reduces the binary image size by a few kilobytes, which makes it worthwhile.
However when LTO is enabled, it makes GCC crash. This is due to a peculiarity
in ChibiOS which has been fixed in later versions. The fix is to remove
the weak attribute tag from the ResetHandler definition on vectors.c
for the ARM Cortex M processors.

Along the way, this exposed a couple of other problems:

1) There was a bug in radioReceive(), where the wrong number of bytes
was being read from the FIFO. This has been fixed.

2) Orchard uses some linker magic to build up the shell command list.
In orchard-shell.h, commands are added to the list by tagging them
with a special attribute keyword that aggregates them into a separate
section. However because the "unused" keyword is also specified, this
causes LTO to optimize away the added shell commands completely.
The definitions for orchard_command() and orchard_command_end() have
been changed from "unused" to "used" to fix this.

The USE_LTO option is now set to "yes" in the Makefile to enable building
with link-time optimization by default.