Fix warnings, turn on -ansi -pedantic
The Makefule now usies -ansi -pedantic when compiling all the code.
This forces strict compliance with the C99 specification. Some cleanups
have been made to silence all the warnings this exposed. A few warnings
remain, but I'm going to hold off on fixing them until I can actually
test the code on hardware to make sure nothing breaks.

Next step will also be to turn on -Wconversion. Testing with this
exposed even more warnings, which will be fixed in the next update.