Add support for the ILI9341 controller on ChibiOS.
This stub connects the controller to the SPI1 channel. We enable the
use of the slave select signal for now since this makes it easy to
use the SPI pins on the J2 header of the Freescale KW019032 board.

We also use GPIO port B pin 1 as the command/data select pin. (This
signal also controls the green LED.)