Enable uGfx widgets
Since we may need more uGfx functionality, support for widgets has been
turned on. This uses up more RAM. Right now we're down to 12032 bytes,
which less than I would have liked.

Enabling widgets also requires the use of an input device, the gtimer
facility and the gevent facility. These all work, but are resource-hungry.
The stack for the gtimer event handler thread was set to 2048 bytes by
default which was too large. This has been reduced to 500 bytes for now.
The event handler table has also been reduced from 32 slots to 4.

Since it doesn't look like we're going to use the I2C or ADC modules
in the KW01, the drivers for these have been turned off to recover a
small amount of memory and code space.