Add SD card support
The ChibiOS MMC driver has been turned on along with uGfx gimage support.
It seems to work, but there's something peculiar about the way the SD
card is detected which I need to sort out. Once in a while the MISO line
seems to get stuck in the low state and the mmcConnect() function spins
forever. I will figure it out when I'm not half asleep.