Add workaround for KW019032 SPI hang
The Freescale/NXP KW019032 board uses Texas Instruments bi-directional
level converters to link the SPI1 channel and GPIO pins to the J1 and J2
headers. With the display/touch/SD module attached, the SPI1_MISO pin
may be driven low at power up, which seems to cause the level converter
to drive the input side low too. This causes all SPI read accesses to
fail since having the pin forced low makes it look like all slaves are
sending 0 bits.

To work around this, we set the PTE0/SPI1_MISO pin to output mode,
set the pin high, and then enable the internal pullup resistor. This
seems to snap the pin back to high state again. The same trick is
also applied to PTE1/SPI1_MOSI just for kicks.

The board now reliably detects the SD card and boots correctly when
the card is not plugged in.