Switch uGFX from BMP to native images
Using the BMP image format adds several KB to the image size. Using
native image format reduces the size significantly and is slightly
faster. The native image format for the ILI9341 display is RGB565.
Some tools have been added in a separate directory to generate image
files in this format.