Switch chip select pin for SD card
The SD card chip select is now port B pin 0, which is also the blue segment
on the tri-color LED on the KW019032 board. This pin is exposed as pin 12
on the J4 header, bit only after a jumper has been soldered (which it
now has been).

Also, the SD card detect message now prints the size of the detected SD card
in megabytes.

Lastly, the makefile has been changed to link with -lc_nano which is
the convention for the ARM embedded toolchain. (If you just build newlib
by itself, you don't get libc_nano.a. This is only done if you use the
build script from the ARM embedded toolchain.)