Fix up the touch screen interface
Currently the screen is set up for landscape mode using a 270-degree rotation.
By default the mouse code tries to apply a transformation to the touch
screen coordinates which it turns out is unnecessary. We need to set a
flag to prevent that. Unfortunately it has to be set in the driver
source itself rather than in the interface header since the structure
where it resides is read-only.

Also, since we're in landscape orientation, the X and Y values returned
from the touch controller chip have to be flipped.