Remove unused early initialization code
The early_init() function has a bunch of extra code in it to initalize
the radio, which is redundant: the radio module already performs a hard
reset of the radio chip when started. Also we are likely not going to
use the I2C port so the early I2C init is not needed either. This saves
a few bytes in ROM.