Always initialize the SD card first
Sometimes, if the touch screen and/or display are initialized first, the
SD card fails to respond. Unfortunately the gfxInit() code always
initializes the disk last, so we do one call to mmc_disk_initialize()
in main() before doing gfxInit(). (When using the ChibiOS MMC code,
we had to connect the SD card in main(), so this was already being
done.)