Make some stabilization fixes
Testing with an animation running shows that the touch controller code
can get stuck when the system is very busy. The problem is that it misses
a byte being read back from the controller, and the SPI code waits
forever for the lost byte to arrive (which it never does). This has been
addressed for now by doing a synchronous read in xptGet().

Also, the PAL code uses a read/modify/write when doing a palSetPad() or
palClearPad(). This is redundant as these regidters are specially designed
not to need a read/modify/writeL bits written as 0 have no effect.

The previous fix also exposed a potential race in the display driver
code. When doing a write_data() or write_index(), we should wait for the
transmit buffer to fully drain rather than just assuming it will happen
in time for us to toggle the data/command pin.

Lastly, the shell thread priority has been raised to be higher than
the main() thread.