Unbreak widget input
With -ansi -pednantic, the compiler complains about anonymous structures,
like the ones in the gwiegetVMT structure definition. I tried to work
around this before, but my workaround actually broke the use of the
MouseUp event, which I couldn't test before because of a broken touch
panel. Now that I have a working panel, the problem has been exposed,
so a new fix is required.

The previous changes to gwin_class.h and gwin_label.c have been reversed
and gwin_class.h and gwin_widget.c have now been updated to deal with
the compiler warning in a way that doesn't break the code.