mplement a custom PWM module for audio
A new module, tpm.c, has been added to generate tones on pin PTC2.

The API is simple:

pwmInit() -- initialize the TPM (called from main())

pwmStartTone (tonefreq) - play a tone where tonefreq is the tone in Hz
pwmStopTone () -- stop plaing a tone

The ChibiOS PWM driver works, but is a little clumsy to use for the
desired application of just generating square wave audio tones. The
tpm.c module is optimized just for that purpose.