Update app-name to use keyboard widget
The app-name app has been updated to use the keyboard widget so that
it's now possible to enter your name from the touch screen. A new
ui-keyboard.c module has been added for this.

Also, some of the orchard-app framework has been repaired. When the
keyboard app exits, it's automatically restarted. Normally it should
be the launcher app that runs, but that hasn't been enabled yet.