Improve the logic in pwmThreadPlay().
Previously, if you queued up a new tune while a current tune was playing,
it was possible for a note to be truncated because chThdResume() would
wake the thread up during the pause between notes instead of when it was
waiting for new work.

Also, it was not possible to queue up the same song twice in a row.

Both of these issues have been fixed.

The new logic is:

- If no tune is currently playing, then a new tune plays.
- If a tune is playing and pwmThreadPlay() is called again with a new
  tune, then the new tune will play when the current tune finishes
- If pwmThreadPlay() is called more than once while a tune is playing,
  only the tune specified in the last call will be played (i.e. tunes
  are not queued).
- pwmThreadPlay(NULL) can still be used to stop a looping tune from
  playing.

Also, the Galaga tune in sound.c has been adjusted slightly to increase
the long note times to make it sound a little more like the original.