Switch to using PEE mode
The Freedom KW019032 board is wired up such that the CLKOUT pin from the
radio is looped back to the EXTAL0 pin (PTA18) on the CPU. This means we
can in fact use the external 32MHz crystal as a timebase for the KW01.

To do this though, we need to include some code in __early_init() to make
sure the radio is programmed to provide the full frequency clock signal.
By default it applies a divisor of 32, which gives us a 1MHz reference
clock, which is too slow. We need a little bit of low-level initialization
to peform a SPI transaction to modify the divisor value in one of the
radio's registers before we can peform the initial clock setup. The
advantage to this should be that we get a more accurate timebase by using
the crystal.