Add support for ugfx events in the orchard-app framework
The orchard-app framework has its own event handling mechanism. This was
probably because the original orchard design didn't make use of uGFX
widgets: instead it relied on irs own input processing. However we have
a touch screen which is implemented as a uGFX mouse device, so we'd like
to be able to process uGFX events too. The problem is that using
geventEventWait() requires us to block inside of an orchard app, which
may prevent us from handling other events, such as reception of radio
packets.

Fortunately, there is a geventRegisterCallback() function which we can
use to channel uGFX events into the orchard app framework. This has been
implemented for the app-launcher app and for the ui-keyboard UI module.

Also:

- app-launcher has been fixed to not leak memory each time it redraws
  the app list

- the ui-keyboard module has been modified to accept a prompt string from
  calling app via the itemlist rather than having the same string hardcoded
  into it all the time. (This allows it to be used by other apps.)

Note that the gevent stack size has also been increased. Sadly, it seems
that the gevent thread may sometimes handle events that require it to
redraw widgets, and the call chain in that case is quite long. Many of
the functions in the chain take many arguments, which leads to a lot of
stack usage.