Fix a couple of bugs:
- When exiting an app, make sure to call geventDetachSource() to disconnect
all of the listener objects. There is a limit of 4 listeners. If you enter
the default badge app and then enter the "Set your name" app, the
ui-keyboard module doesn't get any events because the number of listeners
is already maxed out. Calls to geventDetachSource() have been added to
app-launcher, app-default and ui-keyboard.

- In ui-keyboard.c, don't use gwinSetDefaultStyle(): just set the style
of the keyboard widget instead. Changing the default style affects the way
the up and down arrow buttons are drawn. Previously, when we changed the
default style, if we entered the "Set your name" app and then exited it,
the arrow buttons in the launcher would suddenly have white borders
around them.