Fix the touch/mouse controller
Previously I set up the touch controller interface code to swap the X and Y
axis readings from the XPT2046 chip. This was actually the wrong thing to
do, but I figured out a way to force it to work by setting the flag that
indicates the mouse/touch device can automatically rotate its coordinate
settings. For the default 270-degree rotation, this happened to work.

However we need to fix it so that we can use gdispSetOrientation() works
so that we can flip the screen once we have an orientation sensor, and
in order for that to work correctly, we have to report the native X/Y
axis values and turn off the autorotate flag. This has been corrected
and the calibration values have been regenerated for the new setup.

What I was at it, I also boosted the error correction values to try and
make small widgets respond a little better. I think it helps a little. It
may be adjusted more in the future.

Note that this does not create any currently visible functional change:
it just fixes the behavior of gdispSetOrientation() (which we don't use
yet).