Attempt to be more frugal with RAM
The keyboard widget has some string tables which for some reason always
end up in the data segment for me even though they're declared as const.
Ths consumes about 100 bytes of RAM. To avoid this, the arrays are now
tagged with an attribute tag that forces them into the .rodata section
so that they now end up in ROM.

In the app-name app, avoid using a separate name[] buffer and just update
the config->name field directly. This saves another 16 bytes.

Lastly, some of the apps have been modified to allocate their storage
from the heap rather than relying on local static storage, or the
stack storage from the orchard-app thread. This way, the app storage
is only needed when the app is running.