Reclaim some more RAM, fix a bug in app-setup.c
The ui-keyboard module now dynamically allocates its handles and listener
structures. (The GListener structures are 60 bytes each.) In order to make
this work, a priv member has been added to the OrchardUiContext structure,
similar to what the OrchardApp structure has.

Also, app-setup had a ghLabel3 label which was not used. When I added the
code to dynamically allocate these, I was destroying this handle even
though it wasn't set. By chance, this handle pointer ended up being
0xffff in one my run cases, and it crashed into the Hardfault_Handler()
when trying to free it. The extra handle has been removed.

Lastly, while trying to debug the above bug, I realized that
Hardfault_Handler() wasn't working right. In order to be able to
backtrace through the exception handler, we have to set the sp register
to the process/thread stack. The code there previously was trying to
do this, but in a way which doesn't work with GCC. I tweaked the
assembly a little so that now it updates the stack pointer correctly.
Using 'where' in gdb when we get hard fault works now.