Add a credits app
The credits app scrolls an image sideways across the screen. Generating
the image is a little tricky as it has to be flopped (reversed). I'm
not sure how to correct that in the code. The intent is to make the
image drawing as simple as possible by just blitting the image onto
the screen line by line and scrolling the display as we go.

Sadly the screen can only scroll along its native vertical access. It is
possible to scroll the screen left to right instead of right to left. This
may be needed later if/when we get the screen to rotate with an orientation
sensor.