Try to improve touch accuracy a little more
The touch screen sensor chip has two Z axis readings: Z1 and Z2. I had
difficulty using these originally because the first screen I had was broken
in transit and the touch panel was cracked. The correct way to sense touch
is to read both Z1 and Z1 sensors and check for when the difference
between them decreases. This provides a more accurate touch detection for
some areas of the screen. (The top left corner on mine tends to show a Z1
reading of around 100 which is a bit low.)

The touch driver code now uses both Z1 and Z2 readings to detect a touch,
and the calibration settings have been updated yet again.