Fix firmware updater build with multilib GCC
When using GCC with multilib, we have to force the compiler to select the
right version of libgcc. This means we have to specify the CPU type,
floating point type, and force the use of thumb mode. (The Cortex-M0
only supports 16-bit thumb instructions, not 32-bit ARM instructions.)

Without this, GCC will siliently link the ARM version of libgcc, which
causes a runtime exception when try to use any of the intrinsics routines.