Add some firmware updater fixes
Since the pit and mmc_kinetis_spi code is almost identical to the code
used for the badge, we may as well share those files instead of having
two copies.

Also, if we can't access the SD card or update badge firmware in the
updater, turn on the red LED to signal the user something went wrong.

Now that we can include some of the headers from the badge directory,
some header files in the updater directory are redundant and can be
removed.

The linker script has also been fixed so that it always properly includes
the main() function from the entry.o file. (I found that after modifying
the code in updater.c a little, it wouldn't link this code into the
executable anymore. I don't know the exact behavior that provoked this,
but forcing the inclusion of all .text* sections seems to fix it.)

As a side effect, the updater binary is now slightly smaller.