Implement a proper rand() and save some memory
The newlib libc contains rand() and srand() functions, but we don't want
to use them. Internally, they call the newlib malloc() function, which
doesn't work for us. (We always have to use chHeapAlloc()).

Instead, a rand() and srand() based on the standard C example have been
added. We try to seed with some actual entropy by reading values from
the touch controller sensors, the radio temperature sensor and the
signal strength sensor.

To try to add some more entropy, we also turne on the second PIT timer
(which is otherwise unused) in the __early_init() function and use a few
bits of the timer counter. The amount of time it takes to boot is not
quite fixed as it requires waiting for the SD card and other devices to
come ready.

The amount of entropy is not that great, but it's better than nothing.
At least now we don't always get the same rand() values on every boot.

Also, a few more space-saving measures:

- Only compile in explicit ui modules in the Makefile. Previously we
  were compiling the ui-textentry module, which is unsed (we're using
  the ui-keyboard module instead). This, in conjunction with using a
  local rand() function, saves a few KB of flash space.

- The myImage structure in oled.c was mistakenly marked static, which
  put it in the bss.

- We're not hsing GPIO bank A for external interruprs, so turn it off.
  This saves 16 bytes of RAM since we don't need the porta_channel_map
  structure anymore. (Technically we can turn off portd_channel_map as
  well since we only use a single pin on bank C for interrupts, but
  the interrupts for port C and D are tied to the same pin in the
  KW01, and the ext_lld.c code won't permit disabling one or the other
  without some rewriting..)