Adjust main and gtimer thread stacks
When setting a thread stack size, it's important to use a value that's
a multiple of four. Inside the ChibiOS code, when a new thread is created,
the stack size ends up being used as an offset for an indexed store
instruction. The store is a longword store. If the size is not a multiple
of 4, this results in an attempt to store to an address that is not
longword-aligned, and this triggers an alignment trap.

The gtimer thread needs to be a little bigger now that we've moved
to uGFX 2.6, but not quite as big as 900 bytes. 850 bytes didn't work
because that results in a value that's not a multiple of four. A value
of 828 bytes seems to be ok.

Also, reduce the main thread size back down to 0x300 bytes (it had been
changed to 0x500). We don't actually need to increase the main thread
size: using a bogus gtimer thread size would cause the main thread to
crash when trying to launch the new gtimer thread, but not because it
was running out of stack. (In fact, there was plenty of stack space
left at that point.)