Fix programming of the flash configuration field
The hal_lld.c module contains a _cfm[] block which is supposed to be
written to the flash configuration field at address 0x400. This
controls the default flash security settings. This should be
included in the compiled image, but it was being optmized out. The
volatile keyword has been added to defeat this.

Note that this does not increase the image size since there was
already an empty gap at address 0x400. We're just filling it in now.