Adjust the radio code after some more testing
Now that I have two radio boards again, I can test all the permutations of
frame lengths and filtering settings. The settings should be correct for
all possible configurations now.

Also, make the shout app use the broadcast address as defined in the
radio.h header so that it's correct with hardware filtering on or off.