Make some more radio fixes
The __early_init() routine was not actually resetting the radio correctly.
We have to configure the GPIOE pin 30 pin properly before shitfing its
state will actually work.

Also, turn off the auto-squelch calibrarion for now, in favor of a
hard-coded value. This seems to improve the reliability of RSSI readings.
This may require more experimentation later.