Add some basic code for receiving shouts
The app-notify app now handles received shout messages. They will be
displayed if you're in the launcher of badge apps.

Also, orchard-app now implements support for autoinitializing apps with
the APP_FLAG_AUTOINIT flag set, This allows app-notify to bind itself
to the radio at startup. This works for now, but may be improved later.