Check in tonight's work. Lots of stuff here.
Radio communication in the game should mostly work now, but needs test.

Complete the select, versus, acceptance, and attack/block selection screen
We're currently pretending to be the victim in the attack screens for testing

Proper name display fixed in the battle screens

Attack bitmap operations added to save room

Check in the images, we haven't saved them in awhile.