Upgrade to the latest release of ChibiOS
The current ChibiOS release is 16.1.5 (kernel 3.1.4).

Functionality of the badge is unchanged. There is a small increase in ROM
code size. RAM usage appears unchanged.

One difference that shows up during build is that the BSS size appears
to be 16KB. This is a mirage. It's a side effect of the new linker script
layout. The way the heap space is defined has changed and it how shows
up as a section in the ELF image with the same attributes as the BSS,
so the size(1) command counts it as part of the BSS bytes.