Clean up radio transmit support a little
We want to be able to transmit structures like user info and battle stats
via the radio. We would also like to be able to read them directly out
of the packet payload buffer. In order for that to work, the payload
has to be longword aligned, because the Cortex-M does not support unaligned
accesses, meaning that trying to do a longword load from a memory address
that's not a multiple of 4 will trigger an lignment trap.

In order to achieve this, the protocol field has been extended so that
the header is always a multiple of 4 bytes, and two bytes have been
inserted into the kw01_pkt structure so that the header and payload
start on a longword boundary.

The ping code has also been cleaned up to take advantage of this. This
avoids the need for the ping handler to allocate a user structure on the
stack.