Fix display issues in launcher
The laucher app doesn't quite handle the case where there are more apps
than will fit on the screen correctly. There are two issues:

1) The calculation to determine the number of visible apps fails to take
into account the space taken up by the up/down/go buttons at the bottom
of the screen, which results in the app list being drawn over the top
of those button widgets.

2) We need to always black out the space betweem the end of the app list
and the start of the button widgets. If we have just drawn a full list
of apps, and now we scroll to a partially filled page of apps, the top
of the next page is written with the new apps, but the space following
still displays the apps from the previous page.

Both of these issues have been corrected.