Add chat app, fix 36-point font
The 36-point jupiter pro font was generated with anti-aliasing, and this
make it fairly large. The difference in quality with anti-aliasing is not
enought to justify the added size (about 5K bytes), so it's been
re-generated with anti-aliasing off.

A new chat app has been added, along with a list ui module. This allows
you to select a user and chat with them. You will get a notification when
someone wants to chat with you.