Fix a few nits in the chat routine
The TX buffer will contain a NUL-terminated string, so reduce the string
length provided to the keyboard UI by 1 to prevent it from overrunning
the buffer.

Also, print the user name of a received message in the keyboard entry
screen, a la IRC.

while I'm at it, add some more copyright notices.