Make one more attempt to avoid overruns
In the ui-keyboard module, set the selected field to indicate the total
number of bytes, and use this as the length indicator when sending packets
in app-shout.c and app-chat.c.

Also, change the listUiXX variables in chat_event() to just be ui variables
since they get reused for the keyboard ui as well.