Fix a glitch in the ui-keyboard module
The console widget does not implement backspace. This is a shame, because
we need it (given how hard it can be to type with the virtual keyboard).
Previously there was a hack in place to make it work, but it had one
annoying glitch, which was that if you typed enough to advance to a second
line and then backspaced back to the previous line, a stray copy of the
underbar cursor would be left at position 0 on the second line.

Rather than try to improve the hack, the ui-keyboard module now implements
its own proper backspace routine.