Clean up the chat app a little
Rather than have the app exit and restart when switching from selection
to chat mode, just switch ui modules. (Yes, this is allowed.)

This avoids the need for a couple of global variables which saves a
few bytes of RAM in the data section.

Also, modify the keyboard ui to allow the app to send a ui event to
redraw just the console screen. This avoids having to exit and restart
the screen each time a new message is received.