Update the user search scheme in the chat app
Rather than relying on the enemies list, listen for ping packets and show
them a list the user can choose from. In order to avoid using too much RAM,
we show only the first 50 ping packets that we receive. To store the netid
and username for each possible user takes 20 bytes, which means we could
end up using up to 1000 bytes of RAM. This is a lot, but still within
our limits.