Add support for a second audio voice
The tpm module now supports a second voice thread, using pin PTB1.
This only works if the board.c file is programmed for this, which
currently it's not, as PTB1 is being used by the radio to blink
green when a packet is received. Some discussion is required to
finalize this.

Also, the Galaga tune has been updated with one that's been sampled from
a MIDI file which sounds a little better.