Change gtimer thread to normal priority
When playing music with two voices, we want the music threads not to
be delayed or they'll fall out of sync. The only higher priority thread
is the uGFX timer thread. Reducing its priority to normal doesn't seem
to hurt the touch screen response, so this seems reasonable for now.