Check in a few sound updates and fixes
The sound support has now been changed to reduce the size of the note
structures from 4 bytes to 2. Notes are now specified in MIDI note values
(which range from 0 to 127) and durations are now in milliseconds divided
by 8.

A table of midi notes to frequencies has been added to the tpm module for
conversion. Sadly this table takes 256 bytes of flash, but it allows
tunes to be much smaller, so hopefully it will be worth the savings.

Pin PTB1 is now being programmed as a TPM channel. As a result, the code
in the radio module that blinks the green LED has been turned off. This
pin is at J1 pin 10 on the Freedom board.

Some possibly bogus radio event handler code from orchard-app has been
removed. An event was being triggered on every radio transmit, which
seemed wrong.

Also, the galaga tune now has two voices.