Add a 3rd sound voice and fix a TPM bug
TPM2 channel 0 has been added as a 3rd sound voice, using pin PTB2. This
pin is documented as being used for UART flow control in the Freedom board
user's guide, but it seems ChibiOS doesn't use it for that purpose. There
doesn't seem to be any hardware support for this in the KW01 chip either.
This means it should be available.

Also, a bug in the TPM code has been fixed that allows the use of larger
divisors. We need to reset the timer count value when stopping/starting
the playing of a note in order to make sure that the note plays properly
when we're using large divisor values. If we don't do this, the note
sometimes doesn't play right away. This is less obvious with smaller
divisor values, which is why a value of 8 was originally used. Adding the
third voice seemed to make this problem occur even with a divisor of 8
though, so I was obliged to investigate further.

Resetting the clock count seems to correct the problem. It also allows
the divisor to be set even higher, but doing that may affect frequency
resolution for higher tones, so it's being left alone for now.

Lastly, to exercise the three channels, the galaga tune has been updated
to have 3 voices.