Allow playing music files from SD card
The tpm module now has a pwmFileThreadPlay() function which can be used to
play tunes from the SD card rather than from flash. (This is handy since
flash is small.)

Note that in order for this to work, it was necessary to enable the
_FS_REENTRANT feature in the FatFs code. When testing with the credits
app drawing images and playing music at the same time, some glitching
was observed in the displayed image data. This was due to the fact that
internall the FatFs has a single 512-byte working buffer and both the
Orchard App thread and pwm thread were competing for access to it.
The _FS_REENTRANT feature turns on mutual exclusion protection for
for the working buffer. A syscall.c module has been added which
implements the mutal exclusion routines, using a ChibiOS mutex.

Sadly, because the pwm thread needs to be able to have up to 3 files
open at once in order to play music tracks, it needs to consume a
little more RAM.

A midi directory has been added with a set of sample Super Mario Bros. tune
files. To demo the SD card music feature, these files should be copied to
the SD card before running the credits app.

Two programs have been added for parsing CSV MIDI data and generating
tunes. They are messy and still need work, which will be forthcoming