Add a new credits crawl
Now with more Super Mario!

Also, fix a minor graphics nit in the credits app. Clear the screen
before resetting the scroll setting. If the credits crawl isn't exactly
a multiple of 320 pixels, resetting the hardware scroll offset makes it
look like the screen jumps a little. We clear the screen first to hide
this.