Fix a race condition crash in orchard apps
The orchard app thread(s) run at priority LOWPRIO + 2, which ends up being
4. This is the lowest priority of all threads in the system (except for
the idle thread, which doesn't count). The main thread, which runs the
event handler, is at priority 64, which means it can preempt the orchard
threads. With the chat app, the following can happen:

- You select the chat app -- chat_start() runs
- chat_start() loads the list UI and calls list_start()
- list_start() creates the list widget
- immediately after the list widget is drawn on the screen, but _before_
  the pointer to the widget object is saved in the ListHandles structure,
  a radio packet arrives
- the radio RX interrupt causes the event handler thread (main) to wake
  up, and it calls the orchard app radio callback handler
- the callback handler calls chat_event() which in turn calls list_event()
- list_event() then tries to draw to the list widget using the pointer
  from the ListHandles structure

Because the chat app was interrupted before it could save the list widget
pointer into ListHandles, the p->ghList pointer is invalid. This leads
to a hard fault.

To avoid this, the priority of the orchard app thread has been changed
to NORMALPRIO + 2, which is 66. This places it slightly higher than the
event handler thread, so that it won't be preempted by the event handler
unexpectedly.

Whether or not this is the ideal solution remains to be seen. It might
be better to just have the chat app block preemption while it's initializing
the UI. It does not appear that this change affects the game play mechanics,
however testing needs to be done with many badges to see what happens when
radio traffic increases.