Undo the app thread priority change
Increasing the orchard app thread priority fixes the chat app but breaks
the setup app -- it crashes on exit due to a different kind of race condition.
With the thread running at high priority, the UI event thread can be
preempted, which means a UI event might arrive after the setup_exit()
function is called.

For now, revert the priority change and fix the chat app's problem directly.
To avoid any race conditions, we serialize the creation and destruction of
the app context and list UI via the event handler. Now the whole app is
serialized and there can be no races.