Fix subtle bug in setup app
It turns out there is a specific reason the setup app crashes if you increase
the app thread priority. In the setup_event() handler, when it decides to
exit the app because the user has pressed the OK or CANCEL button, it
calls orchardAppExit(). However unlike all the other apps, it doesn't
immediately afterwards.

What happens with the setup app is an interaction between the app thread,
the main thread and the gtimer thread:

- the gtimer thread triggers a call to setup_event() (user pressed OK)
- setup_event() decodes the event and calls orchardAppExit() and then
  configSave()
- orchardAppExit() causes an event to be dispatched which tells the
  orchard_app_thread() that it should terminate
---- gtimer thread is suspended! ----
---- Setup thread takes over because it has higher priority! ----
- orchard_app_thread() calls setup_exit() handler, which deletes the
  Setup app context
- orchard_app_thread() then dispatches an event which tells the main
  thread to wake up and run its event dispatcher
---- the main thread takes over! ----
- the main thread calls orchardAppRestart(), which creates a new orchard
  app thread running the Launcher app
- The Launcher app draws the first page of the launch list, then goes to
  sleep
---- gtimer thread finally unblocks and gets to run again! ----
- the gtimer thread picks up right where it left off, still inside the
  setup_event() function
- because setup_event() keeps going, it eventually falls through to
  the "//update ui" code which calls gwinSetText() with the p->ghLabelDim
  handle
- the SetupHandles context is no longer valid because it wad free()d a
  long time ago -- as a consequence, the p->ghLabelDim handle is garbage,
  which causes a hard fault in gwinSetText()

To correct this, setup_handler() now returns immediately after calling
orchardAppExit() just like all the other apps. In the OK case, it calls
configSave() first and then calls orchardAppExit() instead of the other
way around (which makes more sense).