Implement software address filtering in radioReceive()
With hardware address filtering disabled, the radio occasionally generates
some spurious RX events. It's not exactly clear why: the interrupt status
registers show that a valud payload is present, but that implies that the
there is a sync byte match and the CRC is good, which is highly unlikely
for bogus radion noise.

This resumts in the default RX handler being called and "No handler for..."
messages being generated on the console.

To avoid this for now, the radioReceive() function now implements some
basic address filtering in software. We check what should be the destination
field in our packet header. If the field is not the broadcast address, and
if it's not the receiving badge's unicast address (netid) then the packet
is assumed to be either bogus or destined for a different badge, so we
ignore it.