Convert the uGFX events to orchard app events.
When an uGFX callback triggers, it now sets off an orchard app event.
Previously, if the gtimer thread initiated a callback, it would
progess up to the current app's event handler, which then in turn
might trigger some other uGFX code. This could lead to call chains
that were 40 stack frames deep in a single thread. This requires the
gtimer thread to be configured with a lot of stack space.

To avoid this, the orchardAppUgfxCallback() function now schedules
an orchard app event. This means that only the code up to the event
callback executes in the gtimer thread, and the rest runs in the
orchard app thread.